Revert "Rename workspacelist to userworkspacelist"
This reverts commit 483388f60a094fd3358da9dab6ef7e2aaa4175a3.
